requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
httpx==0.27.2
python-multipart==0.0.9
starlette==0.36.3 
faiss-cpu
numpy
tiktoken
jinja2
sqlalchemy